hope this does work again

+++ lecturerDataLayer.js
@@ -10,7 +10,6 @@ async function getAlllecturerData() {
     return lecturers;
 
 }
-
 async function addlecturer(full_name, gender, email, phone_number, department, specialization, years_of_experience) {
 
     const connection = getConnection();
@@ -23,14 +22,23 @@ async function addlecturer(full_name, gender, email, phone_number, department, s
 
 async function deletelecturer(lecturerId) {
     const connection = getConnection();
+    const checkQuery = `SELECT * FROM Majors WHERE lecturer_id = ?`;
+    const [involved] = await connection.execute(checkQuery, [lecturerId]);
+    if (involved.length > 0) {
+        return {
+            'success': false,
+            'message': "Unable to delete because the customer is in a sales relationship of an employee"
+        }
+    }
     const query = `DELETE FROM lecturer WHERE lecturer_id = ?`;
     await connection.execute(query, [lecturerId]);
+  
     return {
         'success': true,
         'message': 'lecturer has been deleted'
     }
 }
-
+ 
 async function updatelecturer(lecturerId, updatedData) {
     const connection = getConnection();
     const { full_name, gender, email, phone_number, department, specialization, years_of_experience } = updatedData;
@@ -70,4 +78,4 @@ async function updatelecturer(lecturerId, updatedData) {
 
 }
 
-module.exports = { getAlllecturerData, addlecturer, deletelecturer,updatelecturer };
+module.exports = { getAlllecturerData, addlecturer, deletelecturer, updatelecturer };
